microservices/auto_processing_datasets/requirements.txt
pandas>=2.0.0
pika>=1.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
ollama>=0.6.0
celery[librabbitmq]>=5.3.0
kombu>=5.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
